Move discord_link to data.js
