--- discord_link.js
@@ -1,1 +1,0 @@
-export const discord_link = "https://discord.gg/fqBrGAmavE"
